[jump-to-line addon] Clean up
Issue #3682
