+++ demo/search.html
@@ -14,7 +14,7 @@
 <script src="../addon/search/search.js"></script>
 <script src="../addon/scroll/annotatescrollbar.js"></script>
 <script src="../addon/search/matchesonscrollbar.js"></script>
-<script src="../addon/search/jumpToLine.js"></script>
+<script src="../addon/search/jump-to-line.js"></script>
 <style type="text/css">
       .CodeMirror {border-top: 1px solid black; border-bottom: 1px solid black;}
       dt {font-family: monospace; color: #666;}
@@ -86,7 +86,7 @@ var editor = CodeMirror.fromTextArea(document.getElementById("code"), {
       <dt>Shift-Ctrl-F / Cmd-Option-F</dt><dd>Replace</dd>
       <dt>Shift-Ctrl-R / Shift-Cmd-Option-F</dt><dd>Replace all</dd>
       <dt>Alt-F</dt><dd>Persistent search (dialog doesn't autoclose,
-      enter to find next, shift-enter to find previous)</dd>
+      enter to find next, Shift-Enter to find previous)</dd>
       <dt>Alt-G</dt><dd>Jump to line</dd>
     </dl>
     <p>Searching is enabled by
