+++ doc/compress.html
@@ -243,6 +243,7 @@
           <option value="http://codemirror.net/addon/hint/javascript-hint.js">javascript-hint.js</option>
           <option value="http://codemirror.net/addon/lint/javascript-lint.js">javascript-lint.js</option>
           <option value="http://codemirror.net/addon/lint/json-lint.js">json-lint.js</option>
+          <option value="http://codemirror.net/addon/search/jump-to-line.js">jump-to-line.js</option>
           <option value="http://codemirror.net/addon/lint/lint.js">lint.js</option>
           <option value="http://codemirror.net/addon/mode/loadmode.js">loadmode.js</option>
           <option value="http://codemirror.net/addon/fold/markdown-fold.js">markdown-fold.js</option>
