+++ doc/manual.html
@@ -2234,8 +2234,8 @@ editor.setOption("extraKeys", {
       of <a href="#addon_dialog"><code>openDialog</code></a> when
       available to make prompting for search queries less ugly.</dd>
 
-      <dt id="addon_jumptoline"><a href="../addon/search/jumpToLine.js"><code>search/jumpToLine.js</code></a></dt>
-      <dd>Implements jumpToLine command and binding <code>Alt-G</code> to it.
+      <dt id="addon_jump-to-line"><a href="../addon/search/jump-to-line.js"><code>search/jump-to-line.js</code></a></dt>
+      <dd>Implements a <code>jumpToLine</code> command and binding <code>Alt-G</code> to it.
       Accepts <code>linenumber</code>, <code>+/-linenumber</code>, <code>line:char</code>,
       <code>scroll%</code> and <code>:linenumber</code> formats.
       This will make use of <a href="#addon_dialog"><code>openDialog</code></a>
